Bus annotations now update without removing themselves
This means that if a user selects one, it won't deselect when the bus
updates.

To do so, I made the busesOnRoute property into a dictionary [String:
Bus] so that we could get the bus by busId. I also added a property to
the BusAnnotation for busId, so now we can grab the bus we need for a
specific annotation, update the coordinate proptery and then re-add to
map.

IMPORTANT: Re-adding to map does not create a bunch of duplicates
because behind the scenes, MapKit sees you're adding an annotation that
already exists and just updates it.