Added Split tablview/mapview example
It is in the story board. If you change the segue by ctrl-click and
dragging on the navigationController to the new one, you can select it
as a segue and try it out. It is still rudimentary but it works and I
don't think it will be that hard to adapt it.